requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
youtube-transcript-api>=0.6.0
yt-dlp>=2023.9.24
torch>=2.0.0